run things concurrently + re-organize
