+++ node.go
@@ -1,0 +1,22 @@
+/*
+Package kshaka is a pure Go implementation of the CASPaxos consensus protocol.
+It's name is derived from the Kenyan hip hop group, Kalamashaka.
+
+"CASPaxos is a replicated state machine (RSM) protocol. Unlike Raft and Multi-Paxos,
+it doesn’t use leader election and log replication, thus avoiding associated complexity.
+Its symmetric peer-to-peer approach achieves optimal commit latency in wide-area networks
+and doesn’t cause transient unavailability when any [N−1] of N nodes crash." - [The CASPaxos whitepaper](https://github.com/rystsov/caspaxos/blob/master/latex/caspaxos.pdf)
+
+TODO: add system design here.
+*/
+package kshaka
+
+// Node represents an entity that is both a Proposer and an Acceptor.
+// This is the thing that users who depend on this library will be creating and interacting with.
+// A node is typically a server but it can also represent anything else you want.
+// TODO
+type Node struct {
+	id uint64
+	proposer
+	acceptor
+}
